refactor: move configuration specific cli actions to own module
